requirements.txt
torch~=2.5.0
tensorflow~=2.18.0
tensorboard~=2.18.0
stable-baselines3~=2.4.0a10
sb3-contrib~=2.4.0a10
pandas~=2.2.3
numpy~=1.26.4
gymnasium~=0.29.1
matplotlib~=3.9.2
glfw~=2.7.0
mujoco~=3.2.4
cloudpickle~=3.1.0
moviepy~=1.0.3
ffmpeg~=1.4
pyyaml~=6.0.2
tqdm~=4.66.6
rich~=13.9.3
tensorboard-reducer~=0.3.1
pettingzoo~=1.24.3